reworked factorio mod portal get details about a mod
