+++ src/mod_portal_handler.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"github.com/gorilla/mux"
 	"log"
 	"net/http"
 )
@@ -17,10 +18,37 @@ func FactorioModPortalListMods(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json;charset=UTF-8")
 
 	var statusCode int
-	resp, err, statusCode = listModPortal()
+	resp, err, statusCode = modPortalList()
 
 	if err != nil {
 		resp = fmt.Sprintf("Error in listing mods from mod portal: %s", err)
+		log.Println(resp)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(statusCode)
+}
+
+// FactorioModPortalModInfo returns JSON response with the mod details
+func FactorioModPortalModInfo(w http.ResponseWriter, r *http.Request) {
+	var err error
+	var resp interface{}
+
+	defer func() {
+		WriteResponse(w, resp)
+	}()
+
+	w.Header().Set("Content-Type", "application/json;charset=UTF-8")
+
+	vars := mux.Vars(r)
+	modId := vars["mod"]
+
+	var statusCode int
+	resp, err, statusCode = modPortalModDetails(modId)
+
+	if err != nil {
+		resp = fmt.Sprintf("Error in getting mod details from mod portal: %s", err)
 		log.Println(resp)
 		w.WriteHeader(http.StatusInternalServerError)
 		return
